prepend default test database path with localhost:
having it without the host part requires that either the user has
super-server, in which case `isql` adds that automatically, or that the
user is member of the `firebird` group, so that `isql` (linked with
libfbembed in classic) can access the shared firebird resources (lock
files, password database).

explicitly adding `localhost:` to the path makes requests always go
through the server, regardless it being super or classic, and without
the need for the testing user to be member of the `firebird` group.