move all subs in a separate module and check for $EMBEDDED
the idea is that the same Makefile.PL is used for the main build and
for the build of DBD::FirebordEmbedded in embed/. They are almost exact copies,
with the notable difference of the $EMBEDDED value. It is false for
the main build and true for the embedded build.

having the subs in a separate module avoids a bunch of "subroutine
redefined" warnings when MakeMaker loads embed/Makefile.PL via
require