add Cygwin support
Add a fixup to dbdimp.h to temporarily define _WIN32 before including
ibase.h so that functions are declared as stdcall for calling into the
DLL.

Add a setup_for_cygwin function to inc/FirebirdMaker.pm that finds the
fbclient.dll and links to it.

Add code to Makefile.PL to call setup_for_cygwin for cygwin and
locate_firebird_ms to look in the registry for cygwin as well.